scripts/requirements.txt
### gitlab-project settings
certifi==2020.6.20
chardet==3.0.4
idna==2.10
python-gitlab==2.5.0
requests==2.24.0
urllib3==1.25.10

### hardening_manifest_yaml
# python-gitlab
GitPython
jsonschema
